Update sudoku_generation.h
The sudoku game can have a one or multiple valid solutions and the way of randomization only to generate a valid sudoku may generate not valid one. So after generating a random sudoku I check solving it using backtracking if it solved that means it has at least one solution.